pyproject.toml
[project]
name = "tech-post-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.9",
    "langchain-core>=0.3.59",
    "langchain-openai",
    "langgraph>=0.4.2",
    "langsmith[pytest]>=0.3.4",
    "pytest",
    "langgraph-cli[inmem]",
    "python-dotenv",
    "rich",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
    "streamlit>=1.49.1",
    "faiss-cpu>=1.12.0",
    "langchain-community>=0.3.29",
]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tech_post_agent"]

[tool.setuptools.package-dir]
"tech_post_agent" = "src/tech_post_agent"


